added transform_quanti in groupedlistdisc
