+++ AutoCarver/discretizers/utils/quantitative_discretizers.py
@@ -4,34 +4,53 @@ for a binary classification model.
 
 from typing import Any, Dict, List
 
-from .base_discretizers import GroupedList
-from numpy import array, inf, linspace, nan, quantile, select
-from pandas import DataFrame, Series, isna, notna, unique
-from sklearn.base import BaseEstimator, TransformerMixin
+from .base_discretizers import GroupedList, applied_to_dict_list, GroupedListDiscretizer
+from numpy import array, linspace, nan, quantile
+from pandas import DataFrame, Series, isna, notna
 
-# TODO: issue with ranking of values? -> displayed not in an ascending way
-class QuantileDiscretizer(BaseEstimator, TransformerMixin):
+class QuantileDiscretizer(GroupedListDiscretizer):
     """Builds per-feature buckets of quantiles"""
 
     def __init__(
         self,
         features: List[str],
-        q: int,  # TODO: change this for min_freq
+        min_freq: float,
         *,
         values_orders: Dict[str, Any] = None,
+        str_nan: str = '__NAN__',
         copy: bool = False,
         verbose: bool = False,
     ) -> None:
-        """Discretizes quantitative features into groups of q quantiles"""
+        """Discretizes quantitative features into groups of q quantiles
+
+        Parameters
+        ----------
+        features : List[str]
+            _description_
+        min_freq : float
+            _description_
+        values_orders : Dict[str, Any], optional
+            _description_, by default None
+        str_nan : str, optional
+            _description_, by default '__NAN__'
+        copy : bool, optional
+            _description_, by default False
+        verbose : bool, optional
+            _description_, by default False
+        """
 
         self.features = features[:]
-        self.q = q
+        self.min_freq = min_freq
+
         if values_orders is None:
             values_orders = {}
         self.values_orders = {k: GroupedList(v) for k, v in values_orders.items()}
-        self.quantiles: Dict[str, Any] = {}
+        self.str_nan = str_nan
+
         self.copy = copy
         self.verbose = verbose
+
+        self.q = round(1 / min_freq)  # number of quantiles
 
     def fit(self, X: DataFrame, y: Series = None) -> None:
         """_summary_
@@ -43,83 +62,36 @@ class QuantileDiscretizer(BaseEstimator, TransformerMixin):
         y : Series, optional
             _description_, by default None
         """
+        if self.verbose:  # verbose if requested
+            print(f" - [QuantileDiscretizer] Fit {', '.join(self.features)}")
+
         # computing quantiles for the feature
-        self.quantiles = X[self.features].apply(find_quantiles, q=self.q, axis=0)
+        quantiles = applied_to_dict_list(X[self.features].apply(find_quantiles, q=self.q, axis=0))
 
-        # case when only one feature is discretized
-        if len(self.features) == 1:
-            self.quantiles = {self.features[0]: list(self.quantiles.get(self.features[0]).values)}
+        # storing ordering
+        for feature in self.features:
+            # Converting to a groupedlist 
+            order = GroupedList(quantiles[feature])
 
-        # building string of values to be displayed
-        values: List[str] = []
-        for n, feature in enumerate(self.features):
-            # verbose
-            if self.verbose:
-                print(f" - [QuantileDiscretizer] Fit {feature} ({n+1}/{len(self.features)})")
+            # adding NANs if ther are any
+            if any(isna(X[feature])):
+                order.append(self.str_nan)
 
-            # quantiles as strings
-            feature_quantiles = self.quantiles.get(feature)
-            str_values = ["<= " + q for q in format_list(feature_quantiles)]
+            # storing into the values_orders
+            self.values_orders.update({feature: order})
 
-            # case when there is only one value
-            if len(feature_quantiles) == 0:
-                str_values = ["non-nan"]
-
-            # last quantile is between the last value and inf
-            else:
-                str_values = str_values + [str_values[-1].replace("<= ", ">  ")]
-            values += [str_values]
-
-        # adding inf for the last quantiles
-        self.quantiles = {f: q + [inf] for f, q in self.quantiles.items()}
-
-        # creating the values orders based on quantiles
-        self.values_orders.update(
-            {feature: GroupedList(str_values) for feature, str_values in zip(self.features, values)}
+        # discretizing features based on each feature's values_order
+        super().__init__(
+            self.features,
+            self.values_orders,
+            copy=self.copy,
+            input_dtype='float',
+            output_dtype='str',
+            str_nan=self.str_nan,
         )
+        super().fit(X, y)
 
         return self
-
-    def transform(self, X: DataFrame, y: Series = None) -> DataFrame:
-        """_summary_
-
-        Parameters
-        ----------
-        X : DataFrame
-            _description_
-        y : Series, optional
-            _description_, by default None
-
-        Returns
-        -------
-        DataFrame
-            _description_
-        """
-        # copying dataset if requested
-        Xc = X
-        if self.copy:
-            Xc = X.copy()
-
-        # iterating over each feature
-        for n, feature in enumerate(self.features):
-            # verbose if requested
-            if self.verbose:
-                print(f" - [QuantileDiscretizer] Transform {feature} ({n+1}/{len(self.features)})")
-
-            nans = isna(Xc[feature])  # keeping track of nans
-
-            # grouping values inside quantiles
-            to_input = [
-                Xc[feature] <= q for q in self.quantiles.get(feature)
-            ]  # values that will be imputed
-            values = [[v] * len(X) for v in self.values_orders.get(feature)]  # new values to imput
-            Xc[feature] = select(to_input, values, default=Xc[feature])  # grouping modalities
-
-            # adding back nans
-            if any(nans):
-                Xc.loc[nans, feature] = nan
-
-        return Xc
 
 
 def find_quantiles(
@@ -198,61 +170,3 @@ def find_quantiles(
         quantiles += [max(df_feature.values)]
 
     return quantiles
-
-
-def format_list(a_list: List[float]) -> List[str]:
-    """Formats a list of floats to a list of unique rounded strings of floats"""
-
-    # finding the closest power of thousands for each element
-    closest_powers = [
-        next((k for k in range(-3, 4) if abs(elt) / 100 // 1000 ** (k) < 10)) for elt in a_list
-    ]
-
-    # rounding elements to the closest power of thousands
-    rounded_to_powers = [elt / 1000 ** (k) for elt, k in zip(a_list, closest_powers)]
-
-    # computing optimal decimal per unique power of thousands
-    optimal_decimals: Dict[str, int] = {}
-    for power in unique(closest_powers):  # iterating over each power of thousands found
-        # filtering on the specific power of thousands
-        sub_array = array([elt for elt, p in zip(rounded_to_powers, closest_powers) if power == p])
-
-        # number of distinct values
-        n_uniques = sub_array.shape[0]
-
-        # computing the first rounding decimal that allows for distinction of
-        # each values when rounded, by default None (no rounding)
-        optimal_decimal = next(
-            (k for k in range(1, 10) if len(unique(sub_array.round(k))) == n_uniques),
-            None,
-        )
-
-        # storing in the dict
-        optimal_decimals.update({power: optimal_decimal})
-
-    # rounding according to each optimal_decimal
-    rounded_list: List[float] = []
-    for elt, power in zip(rounded_to_powers, closest_powers):
-        # rounding each element
-        rounded = elt  # by default None (no rounding)
-        optimal_decimal = optimal_decimals.get(power)
-        if optimal_decimal:  # checking that the optimal decimal exists
-            rounded = round(elt, optimal_decimal)
-
-        # adding the rounded number to the list
-        rounded_list += [rounded]
-
-    # dict of suffixes per power of thousands
-    suffixes = {-3: "n", -2: "mi", -1: "m", 0: "", 1: "K", 2: "M", 3: "B"}
-
-    # adding the suffixes
-    formatted_list = [
-        f"{elt: 3.3f}{suffixes[power]}" for elt, power in zip(rounded_list, closest_powers)
-    ]
-
-    # keeping zeros
-    formatted_list = [
-        rounded if raw != 0 else f"{raw: 3.3f}" for rounded, raw in zip(formatted_list, a_list)
-    ]
-
-    return formatted_list
